+++ src/chess/zobrist.rs
@@ -85,6 +85,8 @@ pub fn init() {
 }
 
 pub fn hash(game: &Game) -> ZobristHash {
+    use Player::*;
+
     debug_assert!(
         unsafe { components::SIDE_TO_PLAY != 0 },
         "Zobrist components were not initialised"
@@ -94,52 +96,52 @@ pub fn hash(game: &Game) -> ZobristHash {
 
     // Add piece components to hash
     // White
-    for s in game.board.white_pieces().pawns() {
+    for s in game.board.pawns(White) {
         hash ^= piece_on_square(Player::White, PieceKind::Pawn, s);
     }
 
-    for s in game.board.white_pieces().knights() {
+    for s in game.board.knights(White) {
         hash ^= piece_on_square(Player::White, PieceKind::Knight, s);
     }
 
-    for s in game.board.white_pieces().bishops() {
+    for s in game.board.bishops(White) {
         hash ^= piece_on_square(Player::White, PieceKind::Bishop, s);
     }
 
-    for s in game.board.white_pieces().rooks() {
+    for s in game.board.rooks(White) {
         hash ^= piece_on_square(Player::White, PieceKind::Rook, s);
     }
 
-    for s in game.board.white_pieces().queens() {
+    for s in game.board.queens(White) {
         hash ^= piece_on_square(Player::White, PieceKind::Queen, s);
     }
 
-    for s in game.board.white_pieces().king() {
+    for s in game.board.king(White) {
         hash ^= piece_on_square(Player::White, PieceKind::King, s);
     }
 
     // Black
-    for s in game.board.black_pieces().pawns() {
+    for s in game.board.pawns(Black) {
         hash ^= piece_on_square(Player::Black, PieceKind::Pawn, s);
     }
 
-    for s in game.board.black_pieces().knights() {
+    for s in game.board.knights(Black) {
         hash ^= piece_on_square(Player::Black, PieceKind::Knight, s);
     }
 
-    for s in game.board.black_pieces().bishops() {
+    for s in game.board.bishops(Black) {
         hash ^= piece_on_square(Player::Black, PieceKind::Bishop, s);
     }
 
-    for s in game.board.black_pieces().rooks() {
+    for s in game.board.rooks(Black) {
         hash ^= piece_on_square(Player::Black, PieceKind::Rook, s);
     }
 
-    for s in game.board.black_pieces().queens() {
+    for s in game.board.queens(Black) {
         hash ^= piece_on_square(Player::Black, PieceKind::Queen, s);
     }
 
-    for s in game.board.black_pieces().king() {
+    for s in game.board.king(Black) {
         hash ^= piece_on_square(Player::Black, PieceKind::King, s);
     }